Wrap Rolodex page in Suspense wrapper
